+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 # CHANGELOG
 
-## 0.5.6dev
+## 0.6.1dev
+
+## 0.6.0 (2023-02-27)
+
+* [API Change] Drops support for old versions of IPython (removed imports from `IPython.utils.traitlets`)
+* [Feature] Adds `%%config SqlMagic.autopolars = True` ([#138](https://github.com/ploomber/jupysql/issues/138))
+
+## 0.5.6 (2023-02-16)
+
+* [Feature] Shows missing driver package suggestion message ([#124](https://github.com/ploomber/jupysql/issues/124))
 
 ## 0.5.5 (2023-02-08)